vesuvius: share mirror augmentation and zarr 2 label-writing helpers
